+++ Localization/enUS.lua
@@ -8,7 +8,7 @@ addon.Plugins:LoadLocalization(L);
 -- [[ https://legacy.curseforge.com/wow/addons/krowi-achievement-filter/localization ]] --
 -- [[ Everything after this line is automatically generated from CurseForge and is not meant for manual edit - SOURCETOKEN - AUTOGENTOKEN ]] --
 
--- [[ Exported at 2023-09-06 16-56-11 ]] --
+-- [[ Exported at 2023-09-09 15-54-56 ]] --
 L["%c"] = true
 L["%d/%m/%Y %I:%M %p"] = true
 L["%m/%d/%Y %I:%M %p"] = true
@@ -39,6 +39,8 @@ This value can be changed while the achievement window is open and having one of
 (1) FPS drops and stuttering while adjusting the value, NOT during normal addon use]=]
 L["Achievements"] = true
 L["Active events"] = true
+L["Add addon name to world map icon"] = true
+L["Add addon name to world map icon Desc"] = "Add '{addonName}' to the bottom of the tooltip when mousing over the icon."
 L["Add Locale"] = true
 L["Add Locale Desc"] = "Add the current game's locale to the {wowheadLink}."
 L["Add to / Remove from Watch List"] = true
@@ -241,6 +243,7 @@ L["Dragonflight pre-patch - Sandstorms"] = true
 L["Dragonflight pre-patch - Snowstorms"] = true
 L["Dragonflight pre-patch - Thunderstorms"] = true
 L["Dragonriding"] = true
+L["Dreamsurge"] = true
 L["Drops"] = true
 L["Dueler's Guild"] = true
 L["during"] = true
@@ -639,6 +642,7 @@ L["Other"] = true
 L["Other faction"] = true
 L["Other locations"] = true
 L["Other wings"] = true
+L["Outland Cup"] = true
 L["Overflow"] = true
 L["Overwatch"] = true
 L["Pandaren"] = true